requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-socketio==5.4.0
Adafruit_DHT==1.4.0
RPi.GPIO==0.7.0
python-dotenv==0.19.0
gunicorn==20.1.0
eventlet==0.33.0
Flask-CORS==3.0.10 